Add files via upload
A regression analysis is a very popular technique used in analytical problem solving and
many business problems typically will lend themselves to a regression-based approach,
especially when we are trying to predict or understand future outcomes using data that
already exists on consumer behaviour.
Linear regression is an equation for finding relation between one dependent variable and
one or more independent variable. When there is only one variable it is known as linear
regression however when there are more than one variable it is multiple regression. A
regression model is used to understand and quantify cause-effect relationships.